% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stickBreakingMixture.R
\name{delCorrection_SBM}
\alias{delCorrection_SBM}
\title{SBM correction to Generalized Dirichlet parameters}
\usage{
delCorrection_SBM(p_small, p_large, K)
}
\description{
SBM correction to Generalized Dirichlet parameters
}
