% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalTools.R
\name{logsumexp}
\alias{logsumexp}
\title{Log Sum Exp function}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{real vector, the arguments to exp(). If log(sum(y)) is desired, supply x=log(y).}
}
\value{
real number, numerically stable log(sum(exp(x)))
}
\description{
Numerically stable computation of log(sum(exp(x))).
}
\examples{
logsumexp(1:5)
log(sum(exp(1:5)))
}
