% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalTools.R
\name{rDirichlet}
\alias{rDirichlet}
\title{Single draw from Dirichlet distribution}
\usage{
rDirichlet(alpha, logscale = FALSE)
}
\arguments{
\item{alpha}{real vector, shape parameter.}

\item{logscale, }{should probability vector be returned on the log scale?}
}
\value{
real probability vector, single draw from the Dirichlet(alpha) distribution
}
\description{
Single draw from Dirichlet distribution
}
\examples{
rDiriclet(rep(1,5))
}
\keyword{Dirichlet}
