% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stickBreakingMixture.R
\name{rPostSBM}
\alias{rPostSBM}
\title{Sample Conjugte Posterior: Probability vector from multinomial counts under sparse stick-breaking mixture prior}
\usage{
rPostSBM(x, p_small, p_large = 0, eta, gam, del, w_logout = FALSE,
  z_logout = FALSE)
}
\arguments{
\item{x}{real vector, counts in data}
}
\value{
list
}
\description{
Sample Conjugte Posterior: Probability vector from multinomial counts under sparse stick-breaking mixture prior
}
\examples{
x = c(5, 2, 3)
K = length(x)
rPostSBM(x=x, p_small=0.5, p_large=0.0, eta=1.0e3, gam=1.5, del=1.5)
}
