% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stickBreakingMixture.R
\name{rSBM}
\alias{rSBM}
\title{Simulate from sparse stick-breaking mixture prior}
\usage{
rSBM(K, p_small, p_large = 0, eta, gam, del, logout = FALSE,
  logcompute = TRUE)
}
\description{
Simulate from sparse stick-breaking mixture prior
}
\examples{
a = c(5, 2, 3)
K = length(a)
rSBM(K=K, p_small=0.5, p_large=0.0, eta=1.0e3, gam=1.5, del=1.5)
gamdel = shape_Dir2genDir(a) # tapered parameters to mimic Dirichlet special case
rSBM(K=K, p_small=0.5, p_large=0.0, eta=1.0e3, gam=gamdel[,1], del=gamdel[,2]) # tapered to mimic Dirichlet special case
}
