% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stickBreakingMixture.R
\name{shape_Dir2genDir}
\alias{shape_Dir2genDir}
\title{Convert Dirichlet shape parameter vector into Generalized Dirichlet beta parameters}
\usage{
shape_Dir2genDir(a)
}
\description{
Convert Dirichlet shape parameter vector into Generalized Dirichlet beta parameters
}
